adds PhoneModel
